Moved bibliography from converters.js to generators.js
